+++ cashu/mint/ledger/__init__.py
@@ -1,0 +1,3 @@
+from .ledger import Ledger
+
+__all__ = ["Ledger"]
