+++ cashu/mint/protocols.py
@@ -1,8 +1,8 @@
-from typing import Dict, List, Mapping, Protocol
+from typing import Dict, List, Mapping, Optional, Protocol, Tuple
 
-from ..core.base import Method, MintKeyset, Unit
+from ..core.base import Amount, Method, MintKeyset, Unit
 from ..core.crypto.secp import PublicKey
-from ..core.db import Database
+from ..core.db import Connection, Database
 from ..lightning.base import LightningBackend
 from ..mint.crud import LedgerCrud
 from .db.read import DbReadHelper
@@ -37,3 +37,12 @@ class SupportsDb(Protocol):
 
 class SupportsEvents(Protocol):
     events: LedgerEventManager
+
+
+class SupportsWatchdog(Protocol):
+    async def get_unit_balance_and_fees(
+        self,
+        unit: Unit,
+        db: Database,
+        conn: Optional[Connection] = None,
+    ) -> Tuple[Amount, Amount]: ...
